Making use of the qr's new PriorityQueue for pld priority.
At some point, we'll be honoring crawlDelay, and with different crawl delays for various sites,
it will no longer be true that we can implement pld priority with a simple queue. As such, the
introduction of a priority queue was inevitable. Adding exactly such a queue, based on Redis'
sorted set seemed like the reasonble thing to do.