Robots, and test framework reworking.
The testing framework needed a little reworking after actually trying to use it to
make unit tests for attercob. Hopefully not it's a little more applicable. In addition,
this contains a first pass at a mechanism to asynchronously incorporate robots.txt
requests into downpour. The problem with using reppy before was that it uses plain
old urllib2, which does not sit well with the asynchronous nature of Twisted. Instead,
the polite fetcher will check if it has robots.txt for a particular fqdn, and then
check its expiry to see if it should be refetched. If so, instead of popping off a
request for that particular fqdn, it returns a robots.txt request.