Add a little flexibility in the way robots.txt is fetched
The fetching of robots.txt was a little too rigid in the previous
implementation of PoliteFetcher. Basically, it necessetated a very
strict format on the keys for requests. If that were to change, it
would break the robots.txt fetching. Instead, now, it's doing it
correctly, by getting the next request and basing its robots.txt
request off of the url contained therein, and not based on its key.
In addition, added a callback for when an empty list is popped off.