First pass at solving the multi-complete bug.
What I believe to be the problem was the fact that the number of
simultaneous requests to a given domain were being tracked in a
in-memory map. This is fine so long as one process is running, but
in a situation where you have many processes running, there can
be cases where two processes get a request for a given campaign,
one finished, thinks there are no requests in flight, and no requests
remaining, and considers it finished. The second process then
comes along and realizes there are no more requests in flight in
its process and no more requests for the domain, and so it will
attempt to complete it.